test: 캡션 attr bit29 계약을 통합 테스트로 옮기고 doc 귀속 정정
unit tier 게이트는 source 측 단위 테스트 증가를 금지한다
(src/serializer/control.rs::tests 31 -> 32). 신규 계약을 tests/cases/ 의
통합 테스트로 옮기고 manifest 를 재생성한다.

apply_caption_attr_bit 가 pack_common_attr_bits 의 doc 주석과 그 함수
사이에 삽입돼 비트 레이아웃 문서가 엉뚱한 함수에 붙고 clippy
doc_lazy_continuation 이 나던 것도 함께 정정한다(함수를 doc 블록 위로 이동).
