+++ src/serializer/control.rs
@@ -2165,6 +2165,26 @@ fn serialize_common_obj_attr(common: &CommonObjAttr) -> Vec<u8> {
     w.into_bytes()
 }
 
+/// 개체 공통 속성 attr 의 캡션 존재 플래그(bit 29) 를 실제 캡션 유무와 일치시킨다.
+///
+/// 한글 2022 는 CTRL_HEADER 직후에 캡션 LIST_HEADER 가 오는지를 **이 비트로** 판정한다.
+/// 비트와 실제 캡션이 어긋나면(캡션 있는데 0, 없는데 1) 레코드 스트림을 오독해 문서
+/// 전체 개방을 거부한다 — HWP3/HWPX 출처 표에서 캡션은 방출하면서 비트를 안 켜 발생했다.
+fn apply_caption_attr_bit(ctrl_data: &[u8], has_caption: bool) -> Vec<u8> {
+    const CAPTION_ATTR_BIT: u32 = 1 << 29;
+    let mut out = ctrl_data.to_vec();
+    if out.len() >= 4 {
+        let mut attr = u32::from_le_bytes([out[0], out[1], out[2], out[3]]);
+        if has_caption {
+            attr |= CAPTION_ATTR_BIT;
+        } else {
+            attr &= !CAPTION_ATTR_BIT;
+        }
+        out[0..4].copy_from_slice(&attr.to_le_bytes());
+    }
+    out
+}
+
 /// `CommonObjAttr` 의 enum 필드들로부터 attr u32 비트를 합성한다.
 ///
 /// [#4400] `document_core/converters/common_obj_attr_writer.rs` 에서 이 파일로 이동했다
@@ -2191,26 +2211,6 @@ fn serialize_common_obj_attr(common: &CommonObjAttr) -> Vec<u8> {
 /// - bit 20: size protect when VertRelTo is Para
 /// - bit 26: HWPX GenShape storage high bit 후보
 /// - bit 28: HWPX GenShape numbering category high bit 후보
-/// 개체 공통 속성 attr 의 캡션 존재 플래그(bit 29) 를 실제 캡션 유무와 일치시킨다.
-///
-/// 한글 2022 는 CTRL_HEADER 직후에 캡션 LIST_HEADER 가 오는지를 **이 비트로** 판정한다.
-/// 비트와 실제 캡션이 어긋나면(캡션 있는데 0, 없는데 1) 레코드 스트림을 오독해 문서
-/// 전체 개방을 거부한다 — HWP3/HWPX 출처 표에서 캡션은 방출하면서 비트를 안 켜 발생했다.
-fn apply_caption_attr_bit(ctrl_data: &[u8], has_caption: bool) -> Vec<u8> {
-    const CAPTION_ATTR_BIT: u32 = 1 << 29;
-    let mut out = ctrl_data.to_vec();
-    if out.len() >= 4 {
-        let mut attr = u32::from_le_bytes([out[0], out[1], out[2], out[3]]);
-        if has_caption {
-            attr |= CAPTION_ATTR_BIT;
-        } else {
-            attr &= !CAPTION_ATTR_BIT;
-        }
-        out[0..4].copy_from_slice(&attr.to_le_bytes());
-    }
-    out
-}
-
 pub(crate) fn pack_common_attr_bits(common: &CommonObjAttr) -> u32 {
     let mut a: u32 = 0;
     if common.treat_as_char {
